Fix: 함수 분리
