+++ src/components/game/mainboard/InputWord.tsx
@@ -2,8 +2,6 @@ import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import { Send } from '@mui/icons-material';
 import { useRecoilValue } from 'recoil';
-import toast from 'react-hot-toast';
-import search from './searchWord';
 import {
   CurrentRateType,
   PeoplesType,
@@ -12,9 +10,9 @@ import {
   TimeType,
   WordsType,
 } from '../../../types/gameType';
-import timeLimit from './timer/timeLimit';
 import { userState } from '../../../atoms';
 import { updateRate } from '../../../utils/utils';
+import checkWord from './handleSubmit';
 
 type Props = WordsType &
   CurrentRateType &
@@ -78,9 +76,8 @@ export default function InputWord({
 }: Props) {
   const user = useRecoilValue(userState);
 
-  let interval: NodeJS.Timeout;
   useEffect(() => {
-    interval = setInterval(() => {
+    const interval = setInterval(() => {
       setTime((prev) => prev - 1);
     }, 1000);
     if (time === 0) {
@@ -108,65 +105,11 @@ export default function InputWord({
     }
   }, [currentRate]);
 
-  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    const inputElement = (e.target as HTMLFormElement).querySelector('input');
-    const inputValue = inputElement?.value;
-    const existCheck = words?.find((e) => e === inputValue);
-    if (inputValue) {
-      if (words.length === 0) {
-        setWords([...words, inputValue]);
-        setCurrentRate(currentRate + 1);
-        inputElement.value = '';
-        timeLimit(start, setTime);
-        return true;
-      }
-      if (
-        words[words.length - 1].charAt(words[words.length - 1].length - 1) ===
-        inputValue.charAt(0)
-      ) {
-        if (existCheck) {
-          toast.error('이미 사용된 단어입니다');
-          return false;
-        }
-        setWords([...words, inputValue]);
-        setCurrentRate(currentRate + 1);
-        inputElement.value = '';
-        timeLimit(start, setTime);
-      } else {
-        toast.error('끝말이 이어지지 않습니다');
-      }
-    } else {
-      toast.error('단어를 입력해주세요');
-    }
-    return true;
-  };
-
   return (
     <InputBox
       onSubmit={async (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
-        const inputValue = (e.target as HTMLFormElement).querySelector('input')
-          ?.value;
-        if (start) {
-          if (inputValue) {
-            const isRealWord = await search(inputValue);
-            const isPossible =
-              /^(?=(?:.*[a-z]){3,})(?!.*([a-z])\1{2,})[a-z]{3,10}$/.test(
-                inputValue,
-              );
-            if (isPossible) {
-              if (isRealWord) {
-                handleSubmit(e);
-              } else {
-                toast.error('존재하지 않는 단어입니다!');
-              }
-            } else {
-              toast.error('형식에 맞지 않는 입력입니다!');
-            }
-          }
-        } else {
-          toast.error('시작 버튼을 눌러주세요!');
-        }
+        checkWord(e, start, words, setWords, setCurrentRate, setTime);
       }}
     >
       <GameInput type="text" />
