+++ src/components/game/sound/soundPlay.ts
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const soundUrl = {
+  error: '',
+  success: '',
+  start: '',
+};
+
+function soundPlay() {}
